Added merchant view all ratings for vendor page, account for no rating/reviews

+++ app/models/ratings.py
@@ -7,7 +7,7 @@ class Ratings(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     vendor_id = db.Column(db.Integer, db.ForeignKey('users.id'))
     merchant_id = db.Column(db.Integer, db.ForeignKey('users.id'))
-    star_rating = db.Column(db.Integer)
+    star_rating = db.Column(db.Float)
     comment = db.Column(db.Text)
     date_reviewed = db.Column(db.DateTime)
